requirements.txt
streamlit>=1.44.0,<2.0.0
google-genai>=1.12.0,<2.0.0
gspread>=6.1.0,<7.0.0
google-auth>=2.29.0,<3.0.0
